Refactor: delete unused vars & files
