--- src/main/java/com/pbl/tasktoolintegration/monday/legacy/entity/MondayUserItemId.java
@@ -1,14 +1,0 @@
-package com.pbl.tasktoolintegration.monday.legacy.entity;
-
-import java.io.Serializable;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@EqualsAndHashCode
-@NoArgsConstructor
-@Getter
-public class MondayUserItemId implements Serializable {
-    private String mondayUser;
-    private String mondayItem;
-}
