--- src/main/java/com/pbl/tasktoolintegration/monday/legacy/model/GetUserExpiredItemDto.java
@@ -1,17 +1,0 @@
-package com.pbl.tasktoolintegration.monday.legacy.model;
-
-import java.util.Date;
-import java.util.List;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@Builder
-@ToString
-public class GetUserExpiredItemDto {
-    private String username;
-    private Integer totalExpiredItems;
-}
