--- src/main/java/com/pbl/tasktoolintegration/monday/legacy/repository/MondayUpdateRepository.java
@@ -1,8 +1,0 @@
-package com.pbl.tasktoolintegration.monday.legacy.repository;
-
-import com.pbl.tasktoolintegration.monday.legacy.entity.MondayUpdate;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MondayUpdateRepository extends JpaRepository<MondayUpdate, String> {
-
-}
