+++ src/main/java/com/pbl/tasktoolintegration/monday/model/GetUsersAverageResponseTimeDto.java
@@ -1,0 +1,15 @@
+package com.pbl.tasktoolintegration.monday.model;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@Builder
+@ToString
+public class GetUsersAverageResponseTimeDto {
+    private String username;
+    private double averageResponseTime;
+}
